requirements.txt
asgiref==3.7.2
binary==1.0.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.7.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
setuptools==80.9.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
